[TASK] Move RegisterIconToIconFileRector
Resolves: #3056
