+++ config/register-icon-to-icons-file/register-icon-to-icons-file.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Ssch\TYPO3Rector\Rector\v11\v5\RegisterIconToIconFileRector;
+use Ssch\TYPO3Rector\Rector\v11\v4\RegisterIconToIconFileRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->import(__DIR__ . '/../config.php');
