+++ tests/Rector/v11/v4/RegisterIconToIconFileRector/Fixture/Expected/Configuration/Icons.php.inc
