+++ tests/Rector/v11/v4/RegisterIconToIconFileRector/Fixture/my_extension/ext_localconf.php.inc
