+++ tests/Rector/v11/v4/RegisterIconToIconFileRector/config/configured_rule.php
